// src/common/data/enums/ServiceAccountKind.ts
enum ServiceAccountKind {
  S3 = 'S3',
  FAST_BITCOINS = 'FAST_BITCOINS',
  SWAN = 'SWAN',
  RAMP = 'RAMP',
  WHIRLPOOL = 'WHIRLPOOL',
  COLLABORATIVE_CUSTODY = 'COLLABORATIVE_CUSTODY',
  COMMUNITY_ACCOUNT = 'COMMUNITY_ACCOUNT',
  JOINT_ACCOUNT = 'JOINT_ACCOUNT',
  FNF_ACCOUNT = 'FNF_ACCOUNT',
  DEPOSIT_ACCOUNT = 'DEPOSIT_ACCOUNT'
}

export default ServiceAccountKind